[media-library] Remove TS circular dependency
